+++ tidalapi/__init__.py
@@ -19,4 +19,4 @@ from .user import (  # noqa: F401
     User,
 )
 
-__version__ = "0.7.6"
+__version__ = "0.7.7"