+++ tidalapi/request.py
@@ -53,12 +53,15 @@ class Requests(object):
     """A class for handling api requests to TIDAL."""
 
     user_agent: str
+    # Latest error response that can be returned and parsed after request has been completed
+    latest_err_response: requests.Response
 
     def __init__(self, session: "Session"):
         # More Android User-Agents here: https://user-agents.net/browsers/android
         self.user_agent = "Mozilla/5.0 (Linux; Android 12; wv) AppleWebKit/537.36 (KHTML, like Gecko) Version/4.0 Chrome/91.0.4472.114 Safari/537.36"
         self.session = session
         self.config = session.config
+        self.latest_err_response = requests.Response()
 
     def basic_request(
         self,
@@ -67,6 +70,7 @@ class Requests(object):
         params: Optional[Params] = None,
         data: Optional[JsonObj] = None,
         headers: Optional[MutableMapping[str, str]] = None,
+        base_url: Optional[str] = None,
     ) -> requests.Response:
         request_params = {
             "sessionId": self.session.session_id,
@@ -90,7 +94,10 @@ class Requests(object):
             headers["authorization"] = (
                 self.session.token_type + " " + self.session.access_token
             )
-        url = urljoin(self.session.config.api_v1_location, path)
+        if base_url is None:
+            base_url = self.session.config.api_v1_location
+
+        url = urljoin(base_url, path)
         request = self.session.request_session.request(
             method, url, params=request_params, data=data, headers=headers
         )
@@ -111,7 +118,7 @@ class Requests(object):
                 if refreshed:
                     request = self.basic_request(method, url, params, data, headers)
             else:
-                log.warning("HTTP error on %d", request.status_code)
+                log.debug("HTTP error on %d", request.status_code)
                 log.debug("Response text\n%s", request.text)
 
         return request
@@ -123,6 +130,7 @@ class Requests(object):
         params: Optional[Params] = None,
         data: Optional[JsonObj] = None,
         headers: Optional[MutableMapping[str, str]] = None,
+        base_url: Optional[str] = None,
     ) -> requests.Response:
         """Method for tidal requests.
 
@@ -133,18 +141,20 @@ class Requests(object):
         :param params: The parameters you want to supply with the request.
         :param data: The data you want to supply with the request.
         :param headers: The headers you want to include with the request
+        :param base_url: The base url to use for the request
         :return: The json data at specified api endpoint.
         """
 
-        request = self.basic_request(method, path, params, data, headers)
+        request = self.basic_request(method, path, params, data, headers, base_url)
         log.debug("request: %s", request.request.url)
         try:
             request.raise_for_status()
         except Exception as e:
-            log.info("Got exception {}".format(e))
-            log.debug("Response was {}".format(e.response))
+            log.info("Request resulted in exception {}".format(e))
+            self.latest_err_response = request
             if request.content:
-                log.debug("response: %s", json.dumps(request.json(), indent=4))
+                resp = request.json()
+                log.debug("Request response: '%s'", resp["errors"][0]["detail"])
             if request.status_code and request.status_code == 404:
                 raise ObjectNotFound
             elif request.status_code and request.status_code == 429:
@@ -153,6 +163,25 @@ class Requests(object):
                 # raise last error, usually HTTPError
                 raise
         return request
+
+    def get_latest_err_response(self) -> dict:
+        """Get the latest request Response that resulted in an Exception :return: The
+        request Response that resulted in the Exception, returned as a dict An empty
+        dict will be returned, if no response was returned."""
+        if self.latest_err_response.content:
+            return self.latest_err_response.json()
+        else:
+            return {}
+
+    def get_latest_err_response_str(self) -> str:
+        """Get the latest request response message as a string :return: The contents of
+        the (detailed) error response Response, returned as a string An empty str will
+        be returned, if no response was returned."""
+        if self.latest_err_response.content:
+            resp = self.latest_err_response.json()
+            return resp["errors"][0]["detail"]
+        else:
+            return ""
 
     def map_request(
         self,