+++ HISTORY.rst
@@ -2,6 +2,13 @@
 
 History
 =======
+v0.7.7
+------
+* Add method to get detailed request error response if an error occurred during request. - tehkillerbee_
+* Tests: Add tests tests for ISRC, barcode methods and cleanup exception handling. - tehkillerbee_
+* Feat.: Add support to get tracks by ISRC. - tehkillerbee_, M4TH1EU_
+* Feat.: Add support to get albums by Barcode ID (UPC). - tehkillerbee_, M4TH1EU_
+* Feat.: Add support for a custom base url in `request()` and `basic_request()` to use the new openapi. - M4TH1EU_
 
 v0.7.6
 ------
@@ -178,6 +185,7 @@ v0.6.2
 .. _arnesongit: https://github.com/arnesongit
 .. _Jimmyscene: https://github.com/Jimmyscene
 .. _quodrum-glas: https://github.com/quodrum-glas
+.. _M4TH1EU: https://github.com/M4TH1EU